+++ src/main/java/act/job/Every.java
@@ -20,10 +20,7 @@ package act.job;
  * #L%
  */
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * Mark a method to be a job called at fixed duration. The difference between